requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.7
django-filter==25.1
django-cors-headers==4.7.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
drf-yasg==1.21.10
# ghasedak_sms==1.0.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==24.2
phonenumbers==9.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
django-ratelimit==4.1.0